backend/requirements.txt
fastapi
httpx
uvicorn[standard]
python-dotenv
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
PyGithub
tavily-python